Implement the form functionality.
I implemented the form functionality by creating a method inside
AddResources.vue that catches the user's input and send it to another
method. The second method, which is placed in the parent component
(TheResources.vue) is being provided/injected and gets the arguments
from the first method. Then, create a new object inside the method and
unshift it to the storedResources object.