Implement the delete resource functionality
I implemented the delete resource functionality by creating a method
in TheResources.vue component and provide/inject it in the
LearningResource.vue component. The injected method is triggered by
clicking the button and the pass up the resource id (which came by
props). Then, the actual array if spliced to return the array without
the deleted element.
Using the filter approach doesn't work because it creates a new
array and the components keep using the old array provided by the
provide/inject dynamic.